+++ e2e-tests/src/bin/canister_info.rs
@@ -69,6 +69,7 @@ async fn canister_lifecycle() -> Principal {
             freezing_threshold: None,
             reserved_cycles_limit: None,
             log_visibility: None,
+            log_memory_limit: None,
             wasm_memory_limit: None,
             wasm_memory_threshold: None,
             environment_variables: None,